get filter history on load and fix old save logic
The old logic only added old history to the DOM if it didn't include
the new filter.

It should always show the old history, but only add the new filter value
if it isn't there already.

This logic is fixed in this commit by doing the correct checks.

This commit also moves filter rendering logic to a
separate file called renderFunctions.js